Nvim lua rewrite

+++ installdotfiles.sh
@@ -3,7 +3,7 @@
 # Script to make symlinks to all dotfiles stored here
 dir=$(pwd)
 files="bashrc zshrc tmux.conf vimrc"
-configfolders="nvim rofi sway termite waybar"
+configfolders="nvim"
 echo "current directory is ${dir}"
 echo "moving current dotfiles to backup directory and installing new ones"
 
@@ -11,17 +11,22 @@ mkdir -p ./old
 mkdir -p ./old/config
 
 for file in $files; do
-	echo "Moving ${file}"
-	mv ~/.$file $dir/old
-	echo "Creating symlink to ${file} in home directory"
-	ln -s $dir/$file ~/.$file
+  if [[ -e ~/.$file ]]; then
+    echo "Moving ${file}"
+    mv ~/.$file $dir/old
+  fi
+  echo "Creating symlink to ${file} in home directory"
+  ln -s $dir/$file ~/.$file
 done
 
 echo
+mkdir -p ~/.config
 echo "Installing ~/.config files"
 for folder in $configfolders; do
-  echo "Moving ${folder}"
-  mv ~/.config/$folder $dir/old/config
+  if [[ -e ~/.config/$folder ]]; then
+    echo "Moving ${folder}"
+    mv ~/.config/$folder $dir/old/config
+  fi
   echo "Creating symlink to ${folder} in config"
   ln -s $dir/config/$folder ~/.config/$folder
 done
